+++ src/main/java/com/esgi/pa/api/resources/OptionAbonnementResource.java
@@ -3,23 +3,10 @@ package com.esgi.pa.api.resources;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.OK;
 
-import com.esgi.pa.api.dtos.requests.optionAbonnement.CreateOptionAbonnementRequest;
-import com.esgi.pa.api.dtos.requests.optionAbonnement.GetByIdOptionAbonnementRequest;
-import com.esgi.pa.api.dtos.requests.optionAbonnement.UpdateOptionAbonnementRequest;
-import com.esgi.pa.api.dtos.responses.optionAbonnement.GetOptionAbonnementResponse;
-import com.esgi.pa.api.mappers.OptionAbonnementMapper;
-import com.esgi.pa.domain.entities.Intern;
-import com.esgi.pa.domain.entities.OptionAbonnement;
-import com.esgi.pa.domain.entities.Users;
-import com.esgi.pa.domain.exceptions.TechnicalFoundException;
-import com.esgi.pa.domain.exceptions.TechnicalNotFoundException;
-import com.esgi.pa.domain.services.InternService;
-import com.esgi.pa.domain.services.OptionAbonnementService;
-import com.esgi.pa.domain.services.UserService;
-import io.swagger.annotations.Api;
 import java.util.List;
+
 import javax.validation.Valid;
-import lombok.RequiredArgsConstructor;
+
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -30,6 +17,24 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.esgi.pa.api.dtos.requests.optionAbonnement.CreateOptionAbonnementRequest;
+import com.esgi.pa.api.dtos.requests.optionAbonnement.GetByIdOptionAbonnementRequest;
+import com.esgi.pa.api.dtos.requests.optionAbonnement.UpdateOptionAbonnementRequest;
+import com.esgi.pa.api.dtos.responses.optionAbonnement.GetOptionAbonnementResponse;
+import com.esgi.pa.api.mappers.OptionAbonnementMapper;
+import com.esgi.pa.domain.entities.Intern;
+import com.esgi.pa.domain.entities.OptionAbonnement;
+import com.esgi.pa.domain.entities.Users;
+import com.esgi.pa.domain.exceptions.NotAuthorizationRessourceException;
+import com.esgi.pa.domain.exceptions.TechnicalFoundException;
+import com.esgi.pa.domain.exceptions.TechnicalNotFoundException;
+import com.esgi.pa.domain.services.InternService;
+import com.esgi.pa.domain.services.OptionAbonnementService;
+import com.esgi.pa.domain.services.UserService;
+
+import io.swagger.annotations.Api;
+import lombok.RequiredArgsConstructor;
 
 /**
  * Contient les routes des Service Abonnement
@@ -48,11 +53,15 @@ public class OptionAbonnementResource {
   @GetMapping("{id}")
   public List<GetOptionAbonnementResponse> getAllOptionAbonnements(
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return OptionAbonnementMapper.toGetOptionAbonnementResponse(
-      optionAbonnementService.findAll()
+    if (internService.doesExistForUsers(users)) {
+      return OptionAbonnementMapper.toGetOptionAbonnementResponse(
+        optionAbonnementService.findAll()
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -60,22 +69,30 @@ public class OptionAbonnementResource {
   public GetOptionAbonnementResponse getOptionAbonnementById(
     @PathVariable Long id,
     @Valid @RequestBody GetByIdOptionAbonnementRequest request
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return OptionAbonnementMapper.toGetOptionAbonnementResponse(
-      optionAbonnementService.getById(request.id())
+    if (internService.doesExistForUsers(users)) {
+      return OptionAbonnementMapper.toGetOptionAbonnementResponse(
+        optionAbonnementService.getById(request.id())
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
   @GetMapping("actif/{id}")
   public List<GetOptionAbonnementResponse> getOptionAbonnementsActif(
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return OptionAbonnementMapper.toGetOptionAbonnementResponse(
-      optionAbonnementService.findByStatus()
+    if (internService.doesExistForUsers(users)) {
+      return OptionAbonnementMapper.toGetOptionAbonnementResponse(
+        optionAbonnementService.findByStatus()
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -102,20 +119,25 @@ public class OptionAbonnementResource {
   public GetOptionAbonnementResponse update(
     @Valid @RequestBody UpdateOptionAbonnementRequest request,
     @PathVariable Long id
-  ) throws TechnicalFoundException, TechnicalNotFoundException {
+  )
+    throws TechnicalFoundException, TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    OptionAbonnement optionAbonnement1 = optionAbonnementService.getById(
-      request.id()
-    );
-    OptionAbonnement optionAbonnement = optionAbonnementService.update(
-      optionAbonnement1,
-      request.name(),
-      request.status(),
-      request.optionServiceAbonnementRequests()
-    );
-    return OptionAbonnementMapper.toGetOptionAbonnementResponse(
-      optionAbonnement
+    if (internService.doesExistForUsers(users)) {
+      OptionAbonnement optionAbonnement1 = optionAbonnementService.getById(
+        request.id()
+      );
+      OptionAbonnement optionAbonnement = optionAbonnementService.update(
+        optionAbonnement1,
+        request.name(),
+        request.status(),
+        request.optionServiceAbonnementRequests()
+      );
+      return OptionAbonnementMapper.toGetOptionAbonnementResponse(
+        optionAbonnement
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -124,12 +146,16 @@ public class OptionAbonnementResource {
   public void delete(
     @Valid @RequestBody GetByIdOptionAbonnementRequest request,
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    OptionAbonnement optionAbonnement = optionAbonnementService.getById(
-      request.id()
+    if (internService.doesExistForUsers(users)) {
+      OptionAbonnement optionAbonnement = optionAbonnementService.getById(
+        request.id()
+      );
+      optionAbonnementService.delete(optionAbonnement);
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
-    optionAbonnementService.delete(optionAbonnement);
   }
 }
