+++ src/main/java/com/esgi/pa/api/resources/ServiceAbonnementResource.java
@@ -3,10 +3,24 @@ package com.esgi.pa.api.resources;
 import static org.springframework.http.HttpStatus.CREATED;
 import static org.springframework.http.HttpStatus.OK;
 
+import com.esgi.pa.api.dtos.requests.serviceAbonnement.CreateServiceAbonnementRequest;
+import com.esgi.pa.api.dtos.requests.serviceAbonnement.GetByIdServiceAbonnementRequest;
+import com.esgi.pa.api.dtos.requests.serviceAbonnement.UpdateServiceAbonnementRequest;
+import com.esgi.pa.api.dtos.responses.serviceAbonnement.GetServiceAbonnementResponse;
+import com.esgi.pa.api.mappers.ServiceAbonnementMapper;
+import com.esgi.pa.domain.entities.Intern;
+import com.esgi.pa.domain.entities.ServiceAbonnement;
+import com.esgi.pa.domain.entities.Users;
+import com.esgi.pa.domain.exceptions.NotAuthorizationRessourceException;
+import com.esgi.pa.domain.exceptions.TechnicalFoundException;
+import com.esgi.pa.domain.exceptions.TechnicalNotFoundException;
+import com.esgi.pa.domain.services.InternService;
+import com.esgi.pa.domain.services.ServiceAbonnementService;
+import com.esgi.pa.domain.services.UserService;
+import io.swagger.annotations.Api;
 import java.util.List;
-
 import javax.validation.Valid;
-
+import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -17,23 +31,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-
-import com.esgi.pa.api.dtos.requests.serviceAbonnement.CreateServiceAbonnementRequest;
-import com.esgi.pa.api.dtos.requests.serviceAbonnement.GetByIdServiceAbonnementRequest;
-import com.esgi.pa.api.dtos.requests.serviceAbonnement.UpdateServiceAbonnementRequest;
-import com.esgi.pa.api.dtos.responses.serviceAbonnement.GetServiceAbonnementResponse;
-import com.esgi.pa.api.mappers.ServiceAbonnementMapper;
-import com.esgi.pa.domain.entities.Intern;
-import com.esgi.pa.domain.entities.ServiceAbonnement;
-import com.esgi.pa.domain.entities.Users;
-import com.esgi.pa.domain.exceptions.TechnicalFoundException;
-import com.esgi.pa.domain.exceptions.TechnicalNotFoundException;
-import com.esgi.pa.domain.services.InternService;
-import com.esgi.pa.domain.services.ServiceAbonnementService;
-import com.esgi.pa.domain.services.UserService;
-
-import io.swagger.annotations.Api;
-import lombok.RequiredArgsConstructor;
 
 /**
  * Contient les routes des Service Abonnement
@@ -52,11 +49,15 @@ public class ServiceAbonnementResource {
   @GetMapping("{id}")
   public List<GetServiceAbonnementResponse> getAllServiceAbonnements(
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
-      serviceAbonnementService.findAll()
+    if (internService.doesExistForUsers(users)) {
+      return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
+        serviceAbonnementService.findAll()
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -64,22 +65,30 @@ public class ServiceAbonnementResource {
   public GetServiceAbonnementResponse getServiceAbonnementById(
     @PathVariable Long id,
     @Valid @RequestBody GetByIdServiceAbonnementRequest request
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
-      serviceAbonnementService.getById(request.id())
+    if (internService.doesExistForUsers(users)) {
+      return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
+        serviceAbonnementService.getById(request.id())
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
   @GetMapping("actif/{id}")
   public List<GetServiceAbonnementResponse> getServiceAbonnementsActif(
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
-      serviceAbonnementService.findByStatus()
+    if (internService.doesExistForUsers(users)) {
+      return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
+        serviceAbonnementService.findByStatus()
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -107,21 +116,26 @@ public class ServiceAbonnementResource {
   public GetServiceAbonnementResponse update(
     @Valid @RequestBody UpdateServiceAbonnementRequest request,
     @PathVariable Long id
-  ) throws TechnicalFoundException, TechnicalNotFoundException {
+  )
+    throws TechnicalFoundException, TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    ServiceAbonnement serviceAbonnement1 = serviceAbonnementService.getById(
-      request.id()
-    );
-    ServiceAbonnement serviceAbonnement = serviceAbonnementService.update(
-      serviceAbonnement1,
-      request.name(),
-      request.description(),
-      request.status(),
-      request.imgPath()
-    );
-    return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
-      serviceAbonnement
+    if (internService.doesExistForUsers(users)) {
+      ServiceAbonnement serviceAbonnement1 = serviceAbonnementService.getById(
+        request.id()
+      );
+      ServiceAbonnement serviceAbonnement = serviceAbonnementService.update(
+        serviceAbonnement1,
+        request.name(),
+        request.description(),
+        request.status(),
+        request.imgPath()
+      );
+      return ServiceAbonnementMapper.toGetServiceAbonnementResponse(
+        serviceAbonnement
+      );
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
   }
 
@@ -130,12 +144,16 @@ public class ServiceAbonnementResource {
   public void delete(
     @Valid @RequestBody GetByIdServiceAbonnementRequest request,
     @PathVariable Long id
-  ) throws TechnicalNotFoundException {
+  ) throws TechnicalNotFoundException, NotAuthorizationRessourceException {
     Users users = userService.getById(id);
-    Intern intern = internService.getById(users);
-    ServiceAbonnement serviceAbonnement = serviceAbonnementService.getById(
-      request.id()
+    if (internService.doesExistForUsers(users)) {
+      ServiceAbonnement serviceAbonnement = serviceAbonnementService.getById(
+        request.id()
+      );
+      serviceAbonnementService.delete(serviceAbonnement);
+    }
+    throw new NotAuthorizationRessourceException(
+      "Vous n'etes pas authorisé à accéder à cette ressource"
     );
-    serviceAbonnementService.delete(serviceAbonnement);
   }
 }
